yf_daily/requirements.txt
yfinance==0.2.40
pandas==2.2.2
psycopg2-binary==2.9.9
boto3==1.34.162
python-dateutil==2.9.0.post0
pytz==2024.1
